change ResponseError::HadToClearSacrificial to u32 for chunk size (from usize) avoids parsing error on 64bit vs 32bit and 32 bit should largely suffice for single chunks

+++ src/error.rs
@@ -37,7 +37,7 @@ pub enum ResponseError {
     /// failure to complete request for data
     FailedRequestForData(Data),
     /// had to clear Sacrificial Data in order to complete request
-    HadToClearSacrificial(::NameType, usize),
+    HadToClearSacrificial(::NameType, u32),
 }
 
 impl From<CborError> for ResponseError {
